fix: upload avatar bug

+++ shared/utils/file/avatar_util.py
@@ -11,11 +11,12 @@ from shared.utils.dir_utils import ensure_avatar_path
 
 def save_avatar(typ, old_path, new_path, file):
     ensure_avatar_path(typ)
-    if old_path:
+    if old_path and os.path.exists(old_path):
         try:
             os.remove(old_path)
-        except Exception as e:
-            raise e
+        except Exception:
+            # shouldn't raise exception! swallow it!
+            pass
 
     try:
         f = open(new_path, "wb")
